+++ lib/ft_atoi.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_atoi.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: dayeo <dayeo@student.42.fr>                +#+  +:+       +#+        */
+/*   By: rpeavey <rpeavey@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/03 15:26:17 by dayeo             #+#    #+#             */
-/*   Updated: 2025/04/03 15:26:19 by dayeo            ###   ########.fr       */
+/*   Updated: 2025/04/12 20:43:33 by rpeavey          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,9 +17,9 @@
 ** Skips any isspace(3)
 ** No error checking.
 */
-int	ft_atoi(const char *s)
+long	ft_atoi(const char *s)
 {
-	int	num;
+	long	num;
 	int	flag;
 
 	flag = 1;
